logic update
