+++ src/App.vue
@@ -6,7 +6,7 @@
 </template>
 
 <script>
-import HeaderNav from "./components/HeaderNav.vue";
+import HeaderNav from "./components/interfaceElements/HeaderNav.vue";
 
 export default {
   name: "App",
@@ -16,5 +16,4 @@ export default {
 };
 </script>
 
-<style lang="scss">
-</style>
+<style lang="scss"></style>
